+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.0.1
+
+* Fix README rendering on pub.dev
+
 ## 2.0.0
 
 * New widget: `CommingleMoneyField`